event.loop: seperate loop implementations into three files;
